+++ modules_s/registrar/reply.c
@@ -206,7 +206,7 @@ int codes[] = {
  */
 int send_reply(struct sip_msg* _m)
 {
-	int code;
+	long code;
 	char* msg = MSG_200; /* makes gcc shut up */
 
 	struct lump_rpl* p, *ei;
@@ -230,7 +230,7 @@ int send_reply(struct sip_msg* _m)
 	}
 
 	if (sl_reply(_m, (char*)code, msg) == -1) {
-		LOG(L_ERR, "send_reply(): Error while sending %d %s\n", code, msg);
+		LOG(L_ERR, "send_reply(): Error while sending %ld %s\n", code, msg);
 		return -1;
 	} else return 0;	
 }
